refactor: extract part of my profile to UserInfo component
